phase2/backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend for the multi-user todo web application"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.21",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "httpx>=0.27.2",
]